test: split issue #1805 tests into separate file under line limit
The CI check-file-line-limits job failed because tests/test-merge-queue.mjs
grew to 1770 lines after the issue #1805 additions. Move the new
--auto-resolve / clickable-links coverage to its own file so the main
suite stays under the 1500-line cap.
